display finished items to user

+++ app/account/views.py
@@ -276,7 +276,9 @@ def unconfirmed():
 @account.route('/checklist', methods=['GET', 'POST'])
 @login_required
 def checklist():
-    checklist_items = ChecklistItem.query.filter_by(assignee_id=current_user.id, is_checked=False)
+    checklist_items = ChecklistItem.query.filter_by(assignee_id=current_user.id)
+    completed_items = [item for item in checklist_items if item.is_checked]
+    checklist_items = [item for item in checklist_items if not item.is_checked]
     form = AddChecklistItemForm()
     if form.validate_on_submit():
         #add new checklist item to user's account
@@ -287,18 +289,22 @@ def checklist():
         db.session.add(checklist_item)
         db.session.commit()
         return redirect(url_for('account.checklist'))
-    return render_template('account/checklist.html', user=current_user, form=form, checklist=checklist_items)
+    return render_template('account/checklist.html', form=form, checklist=checklist_items, 
+        completed=completed_items)
 
 
 @account.route('/checklist/delete/<int:item_id>', methods=['GET', 'POST'])
 @login_required
 def delete_checklist_item(item_id):
     checklist_item = ChecklistItem.query.filter_by(id=item_id).first()
-    if checklist_item.is_deletable:
-        db.session.delete(checklist_item)
-        db.session.commit()
+    if checklist_item:
+        if checklist_item.is_deletable:
+            db.session.delete(checklist_item)
+            db.session.commit()
+            return redirect(url_for('account.checklist'))
+        flash('You cannot delete this item', 'error')
         return redirect(url_for('account.checklist'))
-    flash('You cannot delete this item', 'error')
+    flash('Item could not be deleted', 'error')
     return redirect(url_for('account.checklist'))
 
 
@@ -306,9 +312,12 @@ def delete_checklist_item(item_id):
 @login_required
 def complete_checklist_item(item_id):
     checklist_item = ChecklistItem.query.filter_by(id=item_id).first()
-    checklist_item.is_checked = True
-    db.session.add(checklist_item)
-    db.session.commit()
+    if checklist_item:
+        checklist_item.is_checked = True
+        db.session.add(checklist_item)
+        db.session.commit()
+        return redirect(url_for('account.checklist'))
+    flash('Item could not be completed', 'error')
     return redirect(url_for('account.checklist'))
 
 
@@ -316,14 +325,17 @@ def complete_checklist_item(item_id):
 @login_required
 def update_checklist_item(item_id):
     item = ChecklistItem.query.filter_by(id=item_id).first()
-    if item.is_deletable:
-        form = EditChecklistItemForm(item_text=item.text)
-        if form.validate_on_submit():
-            #update checklist item's text
-            item.text=form.item_text.data
-            db.session.add(item)
-            db.session.commit()
-            return redirect(url_for('account.checklist'))
-        return render_template('account/update_checklist.html', form=form)
-    flash('You cannot modify this item', 'error')
+    if item:
+        if item.is_deletable:
+            form = EditChecklistItemForm(item_text=item.text)
+            if form.validate_on_submit():
+                #update checklist item's text
+                item.text=form.item_text.data
+                db.session.add(item)
+                db.session.commit()
+                return redirect(url_for('account.checklist'))
+            return render_template('account/update_checklist.html', form=form)
+        flash('You cannot modify this item', 'error')
+        return redirect(url_for('account.checklist'))
+    flash('Item could not be updated', 'error')
     return redirect(url_for('account.checklist'))
